# swagger/swagger.json
swagger: "2.0"
info:
  description: "API Rest desarrollada en node.js - Modulo estudiantes"
  version: "1.0.0"
  title: "API REST ESTUDIANTES"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "godaycn87@outlook.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "estudiante-app.herokuapp.com"
schemes:
  - "http"
  - "https"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
basePath: "/"
...
paths:
...
  /traerMaterias/{id}:
    get:
      tags:
      - "Calificaciones"
      summary: "RPC de traerMaterias"
      description: ""
      operationId: "traerMaterias"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID de docente"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "300":
          description: "Usted no posee materias"
        "400":
          description: "El id no puede estar vacio"
        "500":
          description: "Id invalido"
      security:
        - Bearer: []
  /listadoAlumnosPorMateria:
    post:
      tags:
      - "Calificaciones"
      summary: "RPC de traerMaterias"
      description: ""
      operationId: "listadoAlumnosPorMateria"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "listadoAlumnosPorMateria"
        in: "body"
        description: "Parametros listado por materia"
        required: true
        schema:
            $ref: '#/definitions/listadoAlumnosPorMateria'
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "300":
          description: "No se encontraron alumnos"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []
  /traerAlumnosPorMateriaExamen:
    post:
      tags:
      - "Calificaciones"
      summary: "RPC de traerAlumnosPorMateriaExamen"
      description: ""
      operationId: "traerAlumnosPorMateriaExamen"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "traerAlumnosPorMateriaExamen"
        in: "body"
        description: "Parametros listado por materia"
        required: true
        schema:
          $ref: '#/definitions/traerAlumnosPorMateriaExamen'
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "300":
          description: "No se encontraron alumnos"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []
  /cargaNotasFinales:
    post:
      tags:
      - "Calificaciones"
      summary: "RPC de traerAlumnosPorMateriaExamen"
      description: ""
      operationId: "cargaNotasFinales"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "notas"
        in: "body"
        description: "Parametros listado par nota finales"
        required: true
        schema:
          $ref: '#/definitions/notas'
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "300":
          description: "No se encontraron alumnos"
        "401":
          description: "No se pudo completar la operación"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []        
  /cargaNotasCursada:
    post:
      tags:
      - "Calificaciones"
      summary: "RPC de traerAlumnosPorMateriaExamen"
      description: ""
      operationId: "cargaNotasCursada"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "notas"
        in: "body"
        description: "Parametros listado par nota cursada"
        required: true
        schema:
          $ref: '#/definitions/notas'
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "300":
          description: "No se encontraron alumnos"
        "401":
          description: "No se pudo completar la operación"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []             
  /cargaNotasCursadaDesdeArchivo:
    post:
      tags:
      - "Calificaciones"
      summary: "RPC de cargaNotasCursadaDesdeArchivo"
      description: ""
      operationId: "cargaNotasCursadaDesdeArchivo"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "notas"
        in: "body"
        description: "Parametros listado par nota cursada desde archivo"
        required: true
        schema:
          $ref: '#/definitions/notasArchivo'
      responses:
        "200":
          description: "OK"
        "400":
          description: "El body no puede estar vacio"
        "300":
          description: "No se encontraron alumnos"
        "401":
          description: "No se pudo completar la operación"
        "500":
          description: "datos invalido"
      security:
        - Bearer: []                    
...
definitions:
...
  listadoAlumnosPorMateria:
    type: "object"
    properties:
      idDocente:
        type: "integer"
        format: "int64"
      idMateria:
        type: "integer"
        format: "int64"
    xml:
      name: "listadoAlumnosPorMateria"      
  traerAlumnosPorMateriaExamen:
    type: "object"
    properties:
      idDocente:
        type: "integer"
        format: "int64"
      idMateria:
        type: "integer"
        format: "int64"
    xml:
      name: "traerAlumnosPorMateriaExamen"  
  notas:
    type: "object"
    properties:
      idDocente:
        type: "integer"
        format: "int64"
      idMateria:
        type: "integer"
        format: "int64"
      asistencia:
        type: "integer"
        format: "int64"
      examenesidExamenes:
        type: "integer"
        format: "int64"
      idInscriptosExamen:
        type: "integer"
        format: "int64"
      nota:
        type: "integer"
        format: "int64"
      recordatorio:
        type: "integer"
        format: "int64"
    xml:
      name: "notas"  
  notasArchivo:
    type: "object"
    properties:
      idDocente:
        type: "integer"
        format: "int64"
      idMateria:
        type: "integer"
        format: "int64"
      asistencia:
        type: "integer"
        format: "int64"
      examenesidExamenes:
        type: "integer"
        format: "int64"
      idInscriptosExamen:
        type: "integer"
        format: "int64"
      nota:
        type: "integer"
        format: "int64"
      recordatorio:
        type: "integer"
        format: "int64"
    xml:
      name: "notas"  
...